Move corepackage.go to cx/packages/packages.go Change CorePackages to DefaultPackages
